Semi-stable release (#1)
* dev21w41a

* Added Monkeydo submodule

Added Monkeydo to .gitmodules

* dev21w41b

* Added functional code before manifest and CSS

The JS-part of this demo is now functional to the point where a Monkeydo manifest can be played on the different windows. CSS styling and an actual manifest will still have to be created.

* Create README.md

* Cake

* Update README.md

* Unsynced working example

Everything is cued up and good to go. Now comes the fun part of syncing the lyrics with the song

* Synchronized lyrics with music

So, yeah this totally didn't take 6 hours to do.

* Style changes and log removal

Fixed the scroll direction for lyrics, made the art window's font clamp (and centered) and also removed some loggers.

player.html will redirect to the main page if opened without a hash and will no longer be caught by crawlers that care about robots.txt